move validation logic to builder
