+++ crates/core/src/table/mod.rs
@@ -94,15 +94,9 @@ use arrow_schema::{Field, Schema};
 use strum::IntoEnumIterator;
 use url::Url;
 
-use HudiInternalConfig::SkipConfigValidation;
-use HudiTableConfig::{DropsPartitionFields, TableType, TableVersion};
-use TableTypeValue::CopyOnWrite;
-
-use crate::config::internal::HudiInternalConfig;
-use crate::config::read::HudiReadConfig;
 use crate::config::read::HudiReadConfig::AsOfTimestamp;
+use crate::config::table::HudiTableConfig;
 use crate::config::table::HudiTableConfig::PartitionFields;
-use crate::config::table::{HudiTableConfig, TableTypeValue};
 use crate::config::HudiConfigs;
 use crate::file_group::reader::FileGroupReader;
 use crate::file_group::FileSlice;
@@ -167,46 +161,6 @@ impl Table {
         self.file_system_view
             .storage
             .register_object_store(runtime_env.clone());
-    }
-
-    fn validate_configs(hudi_configs: &HudiConfigs) -> Result<()> {
-        if hudi_configs
-            .get_or_default(SkipConfigValidation)
-            .to::<bool>()
-        {
-            return Ok(());
-        }
-
-        for conf in HudiTableConfig::iter() {
-            hudi_configs.validate(conf)?
-        }
-
-        for conf in HudiReadConfig::iter() {
-            hudi_configs.validate(conf)?
-        }
-
-        // additional validation
-        let table_type = hudi_configs.get(TableType)?.to::<String>();
-        if TableTypeValue::from_str(&table_type)? != CopyOnWrite {
-            return Err(anyhow!("Only support copy-on-write table."));
-        }
-
-        let table_version = hudi_configs.get(TableVersion)?.to::<isize>();
-        if !(5..=6).contains(&table_version) {
-            return Err(anyhow!("Only support table version 5 and 6."));
-        }
-
-        let drops_partition_cols = hudi_configs
-            .get_or_default(DropsPartitionFields)
-            .to::<bool>();
-        if drops_partition_cols {
-            return Err(anyhow!(
-                "Only support when `{}` is disabled",
-                DropsPartitionFields.as_ref()
-            ));
-        }
-
-        Ok(())
     }
 
     /// Get the latest [Schema] of the table.
